requirements.txt
aiohttp
aiofiles
aria2p
async_generator
cowpy
psutil
wget
TgCrypto
cffi
dnspython
emoji
PyPDF2
cryptg
patool
pendulum
gitpython
asyncurban
google-api-python-client
google-auth-oauthlib
googletrans
python-telegram-bot==12.8
google_images_download
gTTS
gTTS-token
hachoir
REPO_LINK
heroku3
httplib2
humanize
lxml
oauth2client
Pillow
psycopg2
psycopg2-binary
pybase64
pylast
pySmartDL
python-barcode
python-dotenv
pytz
qrcode
requests
search-engine-parser
speedtest-cli
sqlalchemy
telethon
telethon-session-sqlalchemy
urbandict
wikipedia
git+https://github.com/johnwmillr/LyricsGenius.git
git+https://github.com/voins/google-images-download.git@2cd68173c961324a2c41c61e8b6f40a49663ce60
fontTools
lyricsgenius
redis
PyDrive2
pyDownload
PyDictionary
beautifulsoup4
youtube-dl
PyGithub
bs4
covid
base
setuptools
selenium
telegraph
python-dateutil
pyfiglet
jotquote
html_telegraph_poster
csvfaker
git+https://github.com/LonamiWebs/Telethon
git+https://github.com/ahupp/python-magic
pytube
pymongo[srv]
youtube_search
justwatch
deezloader
zipfile38
bitlyshortener
validators
geopy
twitter_scraper
pyaztro
jikanpy
nekos.py
wand
lottie
glitch_this
vcsi
cairosvg
colour
scipy
wordcloud
wheel
